Entropy!
The game now adds random tiles when you move! In terms of core
functionality all that is left is to implement a scoring system and a
game over sequence. Yes, that is right, my 2048 is a zombie! It won't
die! Hey, you knew this commit was going to be random, it practically
says so in the name ;)