Start me up!
Now when you start the game, it will create a new game just like how the
real game does it! I also changed the names of the files because they
make more sense now. Finally, the UI is slightly better than before when
you are playing in the command line interface! Still kind of chintzy
though...